[CO-354] Introduce `NetworkMagic`
Building on the previous commit, we add the `NetworkMagic` datatype,
which allows us to discriminate addresses originating from different
networks.

In order to make this commit as simple and easy-to-review as possible,
I stub out the added arguments of type `NetworkMagic` with a dummy
variable: `fixedNM`. This puts of the (fairly involved) work of
threading `NetworkMagic` values from the nearest occurrence of
`ProtocolMagic` to where said values are needed.

The remainder of the work will come in a separate commit.