[CO-354] Add tests for mainnet/staging ProtocolMagic
We want to be sure that mainnet/staging both have `NMMustBeNothing`
as the `requiresNetworkMagic` field of their `ProtocolMagic` value,
which is read from config.

We add a simple test which can be run on the `lib/configuration.yaml`
file and check multiple YAML keys for valid `RequiresNetworkMagic`s.

We also fix the JSON file, which caused the test to fail because it
did not have an explicit `requiresNetworkMagic: NMMustBeNothing` in
it.